👌 rename the 'report' module to 'tracekit'
