+++ packages/core/src/domain/tracekit/index.ts
@@ -1,3 +1,3 @@
 export { Handler, StackTrace, BrowserError } from './types'
 export { computeStackTrace } from './computeStackTrace'
-export { subscribe, unsubscribe } from './report'
+export { subscribe, unsubscribe } from './tracekit'
